requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
tqdm
faiss-cpu
streamlit
langchain-groq
langchain-chroma